Forward port mms features
CM

Mms: Fix delete template warning message (from ics)

Change delete template warning message for be consistent with the alert buttons

Emoji patch from CM9 for the MMS app

Support for the Unicode subset dedicated for emojis, as used on the iPhone, or even the Gmail website.

Emojis are supported system-wide (display-only) since 4.1 (some glyphs are simply integrated in the system truetype font for the corresponding Unicode code-points).
These black-and-white emojis are displayed in the notification bar, courtesy of Google, but they don't look good in the Conversation view of the MMS app.
This patch adds the ability to send emojis back (as in CM9) and replaces them with their colored bitmap version.